refactor: 把Admin相關API請求獨立分出一支檔案
