+++ src/pages/AdminMainPage.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import styled from "styled-components";
 import { useEffect } from "react";
-import { adminDeleteUserTweet, adminGetUserTweets } from "../api/getUserTweets";
+import { adminDeleteUserTweet, adminGetUserTweets } from "../api/getAdminRelated";
 // import jwt from "jwt-decode";
 
 const PageStyled = styled.div`
